server/requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langgraph==0.6.6
SQLAlchemy==2.0.43
starlette==0.47.3
typing_extensions==4.14.1
uvicorn[h2]==0.35.0
aiosqlite==0.20.0
python-dotenv==1.1.1
greenlet==3.2.4
